# Conflicts:
#	src/main/resources/application.properties
#	src/main/resources/static/index.html

+++ src/main/java/com/mbicycle/imarket/services/impl/OrderServiceImpl.java
@@ -1,6 +1,7 @@
 package com.mbicycle.imarket.services.impl;
 
 import com.mbicycle.imarket.beans.entities.Order;
+import com.mbicycle.imarket.beans.entities.Profile;
 import com.mbicycle.imarket.daos.OrderRepository;
 import com.mbicycle.imarket.services.interfaces.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,18 +15,42 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public boolean add(Order order) {
-        orderRepository.save(order);
-        return true;
+        if (findByProfile(order.getProfile()) == null) {
+            orderRepository.save(order);
+        }
+        return findByProfile(order.getProfile()) != null;
     }
 
     @Override
-    public List<Order> getAll(){
+    public List<Order> getAll() {
         return orderRepository.findAll();
     }
 
     @Override
+    public boolean update(Order orderToUpdate) {
+        Order order = orderRepository.findByProfile(orderToUpdate.getProfile());
+        if (order != null) {
+            order.setDateClosed(orderToUpdate.getDateClosed());
+            order.setDateGot(orderToUpdate.getDateGot());
+            order.setDateOpened(orderToUpdate.getDateOpened());
+            order.setDatePaid(orderToUpdate.getDatePaid());
+            order.setDateReady(orderToUpdate.getDateReady());
+            order.setDateSent(orderToUpdate.getDateSent());
+            orderRepository.save(order);
+        }
+        return order.equals(orderToUpdate);
+    }
+
+    @Override
+    public Order findByProfile(Profile profile) {
+        return orderRepository.findByProfile(profile);
+    }
+
+    @Override
     public boolean delete(Order order) {
-        orderRepository.delete(order);
-        return true;
+        if (findByProfile(order.getProfile()) != null) {
+            orderRepository.delete(order);
+        }
+        return findByProfile(order.getProfile()) == null;
     }
 }
